Move trusted setup functions to their own file
