+++ src/common.c
@@ -363,15 +363,14 @@ void compute_powers(fr_t *out, const fr_t *x, uint64_t n) {
 /**
  * Deserialize a blob (array of bytes) into a polynomial (array of field elements).
  *
- * @param[out] p    The output polynomial (array of field elements)
- * @param[in]  blob The blob (an array of bytes)
+ * @param[out]  p           The output polynomial (array of field elements)
+ * @param[in]   blob        The blob (an array of bytes)
+ * @param[in]   num_fields  The number of field elements in the blob
  */
-C_KZG_RET blob_to_polynomial(Polynomial *p, const Blob *blob) {
+C_KZG_RET blob_to_polynomial(fr_t *p, const uint8_t *blob, size_t num_fields) {
     C_KZG_RET ret;
-    for (size_t i = 0; i < FIELD_ELEMENTS_PER_BLOB; i++) {
-        ret = bytes_to_bls_field(
-            &p->evals[i], (Bytes32 *)&blob->bytes[i * BYTES_PER_FIELD_ELEMENT]
-        );
+    for (size_t i = 0; i < num_fields; i++) {
+        ret = bytes_to_bls_field(&p[i], (Bytes32 *)&blob[i * BYTES_PER_FIELD_ELEMENT]);
         if (ret != C_KZG_OK) return ret;
     }
     return C_KZG_OK;
