+++ src/setup.h
@@ -13,10 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 #pragma once
 
 #include "common.h"
 #include "eip4844.h"
 #include "eip7594.h"
-#include "setup.h"
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+// Public Functions
+////////////////////////////////////////////////////////////////////////////////////////////////////
+
+C_KZG_RET load_trusted_setup(
+    KZGSettings *out,
+    const uint8_t *g1_monomial_bytes,
+    size_t num_g1_monomial_bytes,
+    const uint8_t *g1_lagrange_bytes,
+    size_t num_g1_lagrange_bytes,
+    const uint8_t *g2_monomial_bytes,
+    size_t num_g2_monomial_bytes,
+    size_t precompute
+);
+
+C_KZG_RET load_trusted_setup_file(KZGSettings *out, FILE *in, size_t precompute);
+
+void free_trusted_setup(KZGSettings *s);
